move getMatchPath to gatsby-core-utils
