+++ packages/gatsby-core-utils/src/index.ts
@@ -9,5 +9,6 @@ export { getGatsbyVersion } from "./get-gatsby-version"
 export { getTermProgram } from "./get-term-program"
 export { fetchRemoteFile, IFetchRemoteFileOptions } from "./fetch-remote-file"
 export { isTruthy } from "./is-truthy"
+export { getMatchPath } from "./match-path"
 export * from "./service-lock"
 export * from "./site-metadata"
