+++ packages/gatsby-plugin-page-creator/package.json
@@ -29,6 +29,7 @@
     "chokidar": "^3.5.1",
     "fs-exists-cached": "^1.0.0",
     "gatsby-page-utils": "^1.4.0-next.1",
+    "gatsby-core-utils": "^2.4.0-next.1",
     "gatsby-telemetry": "^2.4.0-next.1",
     "globby": "^11.0.2",
     "lodash": "^4.17.21"
