+++ packages/gatsby-plugin-page-creator/src/create-pages-from-collection-builder.ts
@@ -4,7 +4,7 @@ import { createPath } from "gatsby-page-utils"
 import { Reporter } from "gatsby/reporter"
 import { Options as ISlugifyOptions } from "@sindresorhus/slugify"
 import { reverseLookupParams } from "./extract-query"
-import { getMatchPath } from "./get-match-path"
+import { getMatchPath } from "gatsby-core-utils"
 import { getCollectionRouteParams } from "./get-collection-route-params"
 import { derivePath } from "./derive-path"
 import { watchCollectionBuilder } from "./watch-collection-builder"
